Compute various normalization methods on a count matrix

Given a matrix of gene counts and a grouping factor identifying the group membership for each column calculate normalized expression values using a variety of gene expression normalization techniques.

normalize_count_matrix(
  x,
  groups,
  filter = TRUE,
  min_count = 10,
  min_total_count = 15,
  large_n = 10,
  min_prop = 0.7,
  ref_column = NULL,
  log_ratio_trim = 0.3,
  sum_trim = 0.05,
  do_weighting = TRUE,
  acutoff = -1e+10,
  p = 0.75,
  n_sub = 1000,
  fit_type = c("parametric", "local", "mean", "glmGamPoi"),
  batch = NULL,
  norm_factors = NULL,
  window = 0.05
)

Arguments

x
SummarizedExperiment object containing a counts assay with raw counts or a matrix with features as rows and samples as columns.

groups
a group level continuous or categorial covariate associated with each sample or column in the object. The order of groups must match the order of the columns in object.

filter
bool. Default = TRUE. Should the lowly expressed genes be removed prior to normalization? Uses the filterByExpr function in the edgeR package.

min_count
Passed to edgeR::filterByExpr the minimum count required for at least some samples. Default 10.

min_total_count
Passed to edgeR::filterByExpr the minimum total count required. Default 15.

large_n
Passed to edgeR::filterByExpr the number of samples per group that is considered to be "large". Default 10.

min_prop
Passed to edgeR::filterByExpr the minimum proportion of samples in the smallest group that express the gene. Default 0.7.

ref_column
Column to use as reference for method = "TMM". Can be a column number or a numeric vector of length nrow(object). Default = NULL. Used by edgeR::calcNormFactors

log_ratio_trim
the fraction (0 to 0.5) of observations to be trimmed from each tail of the distribution of log-ratios (M-values) before computing the mean. Used by method="TMM" for each pair of samples. Default = 0.3. Used by edgeR::calcNormFactors

sum_trim
the fraction (0 to 0.5) of observations to be trimmed from each tail of the distribution of A-values before computing the mean. Used by method="TMM" for each pair of samples. Default = 0.05. Used by edgeR::calcNormFactors

do_weighting
logical, whether to use (asymptotic binomial precision) weights when computing the mean M-values. Used by method="TMM" for each pair of samples. Default = TRUE. Used by edgeR::calcNormFactors

acutoff
minimum cutoff applied to A-values. Count pairs with lower A-values are ignored. Used by method="TMM" for each pair of samples. Default = -1e10. Used by edgeR::calcNormFactors

p
numeric value between 0 and 1 specifying which quantile of the counts should be used by method="upperquartile". Default = 0.75. Used by edgeR::calcNormFactors

n_sub
the number of genes to subset. Default = 1000. Used by DESeq2::vst

fit_type
either "parametric", "local", "mean", or "glmGamPoi" for the type of fitting of dispersions to the mean intensity. Used by DESeq2::vst. Default = "parametric"

norm_factors
optional normalization scaling factors. Used by qsmooth::qsmooth

window
window size for running median which is a fraction of the number of rows in object. Default is 0.05. Used by qsmooth::qsmooth

batch.
(Optional) batch covariate (multiple batches are not allowed). If batch covariate is provided, Combat() from sva is used prior to qsmooth normalization to remove batch effects. See Combat() for more details. Used by qsmooth::qsmooth

Value

List object with items containing matrices of normalized log count per million values for each normalization method along with the original counts.

Details

The following normalization methods are computed

"TMM" Trimmed Mean of M-values computed by edgeR::calcNormFactors

"RLE" Relative log expression computed by edgeR::calcNormFactors

"UQ" Upper Quartile normalization computed by edgeR::calcNormFactors

"VST" Variance Stabalizing Transformation computed by DESeq2::vst

"RLog" Regularized Log transformation computed by DESeq2::rlog

"QSmooth" Smoothed Quantile Normalization computed by qsmooth::qsmooth

"Total Count" Library Size Normalization only

The function returns a list with the log count per million normalized values for each of the above methods. The log2 scaled original counts are also returned. For all methods except vst and rlog a pseudocount of 2 is added to avoid taking logs of 0.