MD Plot Function

Create an MD plot from the given dataframe

plot_md(
  df,
  x = logCPM,
  y = logFC,
  sig_col = FDR,
  fdr = 0.05,
  lfc = 0,
  annotate_counts = TRUE,
  xmax_label_offset = 0.8,
  ymax_label_offset = 0.5,
  ymin_label_offset = 0.5
)

Arguments

df
dataframe containing log-fold-change, p-value, and logCPM columns.

x
column in dataframe containing the logCPM data. Default (logCPM)

y
column in dataframe containing the log-fold-change values. Default (logFC)

sig_col
column in dataframe containing the results from significance testing. Default (FDR)

fdr
numeric. Significance level cutoff for plotting. Values below the given fdr threshold are considered significant. Default (0.05)

lfc
numeric. Log-fold-change cutoff for plotting. Values greater than the abs(lfc) and less than fdr are displayed as differentially expressed. Default(0)

annotate_counts
TRUE/FALSE. Annotate the plot with the summarized gene counts

xmax_label_offset
numeric. Value between 0 and 1 inclusive. Controls the x-position of the count labels

ymax_label_offset
numeric. Value between 0 and 1 inclusive. Controls the y-position of the 'up' count label

ymin_label_offset
numeric. Value between 0 and 1 inclusive. Controls the y-position of the 'down' count label

Value

ggplot MD plot

Examples

if (FALSE) {
library(edgeR)
library(coriell)

# create some fake data
x <- data.frame(
  ctl1 = rnbinom(1000, size = 0.4, prob = 1e-5),
  ctl2 = rnbinom(1000, size = 0.4, prob = 1e-5),
  trt1 = rnbinom(1000, size = 0.4, prob = 1e-5),
  trt2 = rnbinom(1000, size = 0.4, prob = 1e-5),
  row.names = paste0("gene", 1:1000)
)

# run edger pipeline
group <- factor(c(1, 1, 2, 2))
y <- DGEList(counts = x, group = group)
y <- calcNormFactors(y)
design <- model.matrix(~group)
y <- estimateDisp(y, design)

# To perform quasi-likelihood F-tests:
fit <- glmQLFit(y, design)
qlf <- glmQLFTest(fit, coef = 2)

# convert the results object to a dataframe
res_df <- edger_to_df(qlf)

# Create md plot
plot_md(res_df)
}