Z-score a dataframe

Convenience function for scaling a dataframe by rows or columns.

zscore_df(df, by = c("row", "column"), ...)

Arguments

df

data.frame or matrix of numeric columns

by

a character string indicating which axis to scale by. Either "row" or "column". Default ("row")

...

Additional arguments passed to `scale`

Value

data.frame with scaled values

Examples

# generate example data
df <- data.frame(
  a = runif(10),
  b = runif(10),
  c = runif(10)
)
rownames(df) <- paste0("row", rownames(df))

# z-score each row of the data.frame
zscore_df(df, by = "row")
#>                 a          b           c
#> row1  -1.11913677  0.3132969  0.80583991
#> row2   0.96519691 -1.0315030  0.06630609
#> row3  -1.15061448  0.4912553  0.65935923
#> row4   0.09570775 -1.0444130  0.94870521
#> row5  -1.15129685  0.4989238  0.65237306
#> row6   0.79303279  0.3303428 -1.12337556
#> row7  -0.39469592 -0.7424187  1.13711460
#> row8   1.09316321 -0.2244859 -0.86867732
#> row9   0.65636751  0.4945472 -1.15091474
#> row10  0.63681348 -1.1525838  0.51577037