Perform simple imputation on rows of a matrix

Impute NA values for each row of a matrix.

impute(x, fun = median)

Arguments

x

numeric matrix or data.frame that can be converted to a numeric matrix

fun

Imputation function to apply to rows of the matrix. Default median

Value

Matrix with imputed values

Examples

# Create a matrix of values with NAs
X <- matrix(runif(25), 5, dimnames = list(paste0("CpG", 1:5), paste0("Sample", 1:5)))
X[sample.int(25, 5)] <- NA
X
#>        Sample1   Sample2     Sample3   Sample4   Sample5
#> CpG1 0.1724830 0.4706404 0.379278272 0.6759150 0.7050502
#> CpG2 0.3209764        NA 0.903148378 0.3756179 0.6459561
#> CpG3 0.6101362 0.8233726 0.562247924 0.3213041 0.3567089
#> CpG4        NA 0.8845667 0.004998337        NA 0.7490353
#> CpG5 0.7365768 0.6457064 0.351634632        NA        NA

# Impute missing values with row medians
impute(X)
#>        Sample1   Sample2     Sample3   Sample4   Sample5
#> CpG1 0.1724830 0.4706404 0.379278272 0.6759150 0.7050502
#> CpG2 0.3209764 0.5107870 0.903148378 0.3756179 0.6459561
#> CpG3 0.6101362 0.8233726 0.562247924 0.3213041 0.3567089
#> CpG4 0.7490353 0.8845667 0.004998337 0.7490353 0.7490353
#> CpG5 0.7365768 0.6457064 0.351634632 0.6457064 0.6457064

# Impute missing values with row mins
impute(X, min)
#>          Sample1   Sample2     Sample3     Sample4   Sample5
#> CpG1 0.172483016 0.4706404 0.379278272 0.675914955 0.7050502
#> CpG2 0.320976367 0.3209764 0.903148378 0.375617943 0.6459561
#> CpG3 0.610136199 0.8233726 0.562247924 0.321304126 0.3567089
#> CpG4 0.004998337 0.8845667 0.004998337 0.004998337 0.7490353
#> CpG5 0.736576827 0.6457064 0.351634632 0.351634632 0.3516346