Perform vote counting on differential expression results

Perform vote-counting on a list of differential expression results. This method uses user defined p-value and log fold-change cutoffs to determine the number of times a gene is differentially expressed across multiple studies. A gene is then considered 'meta-differentially expressed' if it is differentially expressed in the user defined proportion of studies.

meta_vote(
  exp_list,
  lfc_col = "logFC",
  pval_col = "FDR",
  gene_col = "feature_id",
  lfc = 0,
  pval = 0.05,
  meta_prop = 0.5,
  all_common = FALSE,
  plot = FALSE
)

Arguments

exp_list

list() of data.frames of differential expression results.

lfc_col

Column name in expression data.frames that contains the log fold-change estimates. default = "logFC".

pval_col

Column name in expression data.frames that contains the significance values. default = "FDR".

gene_col

Column name in the expression data.frames that contains the gene IDs. default = "feature_id".

lfc

Numeric. Log fold-change value used as cutoff for determining differential expression in each study. default = 0.

pval

Numeric. Significance value used as cutoff for determining differential expression in each study

meta_prop

Numeric. Proportion of studies a gene must be differentially expressed in to be considered as 'meta-differentially expressed'. default = 0.5.

all_common

Logical. Use only genes present in all experiments. default = FALSE.

plot

Logical. Display a metavolcano ggplot2 plot of the results. default = FALSE.

Value

data.table of meta-expression results with the following columns:

gene_col: Genes included in the final voting procedure.

n_de: The total number of studies in which the gene was differentially expressed (either up or down-regulated).

prop_de: The proportion of studies in which the gene was differentially expressed (either up or down-regulated).

sign_consistency: The number of studies in which a gene was up-regulated minus the number of studies in which a gene was down-regulated.

vote: The final vote for the gene. based on sign(sign consistency) & abs(sign consistency) >= meta_prop * N studies.

Details

There are certain caveats to consider when using the vote counting method. 1. combining FDR and logFC cutoffs after the fact destroys FDR correction. See McCarthy & Smyth, 2009 for more details. 2. The direction of expression is not considered when the final vote is cast. For example, if geneA is up-regulated in 4 studies but down-regulated in 1 then the final vote for geneA will be 'up-regulated'.