Perform a permutation correlation test on every row of a matrix

Compute a correlation value for every row of X against the vector y and n random permutations of y. If the number of possible permutations is less than the the argument n_perm then an exact test is performed instead. In both cases the function returns a data.frame of the original data with additional columns for the test statistic, empirical p-value, and FDR corrected empirical p-value.

permutation_correlation_test(X, y, n_perm = 10000, n_core = 1, ...)

Arguments

X
numeric matrix or data.frame that can be converted to a numeric matrix

y
numeric vector of values to correlate with rows of X

n_perm
integer. The desired number of permutations to sample from. Default (10,000)

n_core
integer. The number of cores to use for processing. Default (1)

...
Additional arguments to pass to `cor` function

Examples

# generate example data
X <- matrix(runif(1e3 * 10), nrow = 1e3, ncol = 10)
y <- 1:10
dimnames(X) <- list(paste("feature", 1:1e3, sep = "."), paste("sample", 1:10, sep = "."))

# correlate each row of X with 1,000 random permutations of vector y
res <- permutation_correlation_test(X, y, n_perm = 1e3, n_core = 8, method = "spearman")

head(res)

#>            sample.1    sample.2   sample.3  sample.4  sample.5  sample.6
#> feature.1 0.4250915 0.317255655 0.95068810 0.1320049 0.2311153 0.2025191
#> feature.2 0.6494543 0.480898274 0.82278898 0.2116591 0.9115859 0.5900253
#> feature.3 0.1010948 0.306155272 0.07322219 0.4193117 0.5564470 0.4480192
#> feature.4 0.7942241 0.002179572 0.05246251 0.6645289 0.2907057 0.6109989
#> feature.5 0.7925411 0.911908104 0.58865616 0.9881507 0.5871362 0.4362855
#> feature.6 0.7326144 0.442248390 0.29841096 0.7442068 0.5055164 0.7512962
#>             sample.7   sample.8  sample.9 sample.10          cor empirical.p
#> feature.1 0.09381644 0.69755122 0.7567052 0.1181995 -0.248484848       0.217
#> feature.2 0.14177546 0.37434826 0.8033095 0.8115017  0.006060606       0.505
#> feature.3 0.69939071 0.81707889 0.4928809 0.8332423  0.866666667       0.001
#> feature.4 0.04673134 0.03060915 0.8069986 0.4487471  0.078787879       0.397
#> feature.5 0.80862167 0.79165509 0.2006797 0.1929895 -0.636363636       0.020
#> feature.6 0.06345900 0.57620590 0.9729349 0.1154842 -0.030303030       0.458
#>                 FDR
#> feature.1 0.4781705
#> feature.2 0.5065196
#> feature.3 0.1666667
#> feature.4 0.4927361
#> feature.5 0.4367816
#> feature.6 0.4942408