Parallel Implementation of a Correlation Permutation Test

Performs a correlation permutation test over all rows of a data.frame using multiple cores

permutation_correlation_test(
  df,
  y,
  n_cores = 1,
  n_perm = 1000,
  cor_method = "spearman",
  p_adjust_method = "fdr"
)

Arguments

df
data.frame. Site by Samples data.frame

y
numeric vector. Numeric vector of values used to correlate with each row of df

n_cores
numeric. Number of cores to use for parallel processing. Default 1.

n_perm
numeric. Number of permutations to perform on each row. Default 1000.

cor_method
a character string indicating which correlation coefficient (or covariance) is to be computed. One of "pearson", "kendall", or "spearman" (default): can be abbreviated.

p_adjust_method
a character string indicating which of the p.adjust.methods to use for correction. Default "fdr".

Value

df with additional columns for correlations, empirical p-values, and fdr adjusted p-values.

Examples

if (FALSE) {
library(coriell)
library(methylKit)

# define age dataframe -- ages matching column order
ages <- data.frame(age = c(30, 80, 34, 30, 80, 40, 35, 80))

# simulate a methylation dataset
sim_meth <- dataSim(
  replicates = 8,
  sites = 1000,
  treatment = c(rep(1, 4), rep(0, 4)),
  covariates = ages,
  sample.ids = c(paste0("test", 1:4), paste0("ctrl", 1:4))
)

# extract the methylation as percentages and coerce to data.frame
perc_meth <- as.data.frame(percMethylation(sim_meth))

# perform permutation testing using 4 cores and 10000 permutations
res <- permutation_correlation_test(perc_meth,
  y = ages$age,
  n_cores = 4,
  n_perm = 10000,
  cor_method = "spearman",
  p_adjust_method = "fdr"
)
}