Perform a permutation correlation test on every row of a matrix

Compute a correlation value for every row of X against the vector y and n random permutations of y. If the number of possible permutations is less than the the argument n_perm then an exact test is performed instead. In both cases the function returns a data.frame of the original data with additional columns for the test statistic, empirical p-value, and FDR corrected empirical p-value.

permutation_correlation_test(X, y, n_perm = 10000, n_core = 1, ...)

Arguments

X
numeric matrix or data.frame that can be converted to a numeric matrix

y
numeric vector of values to correlate with rows of X

n_perm
integer. The desired number of permutations to sample from. Default (10,000)

n_core
integer. The number of cores to use for processing. Default (1)

...
Additional arguments to pass to `cor` function

Examples

# generate example data
X <- matrix(runif(1e3 * 10), nrow = 1e3, ncol = 10)
y <- 1:10
dimnames(X) <- list(paste("feature", 1:1e3, sep = "."), paste("sample", 1:10, sep = "."))

# correlate each row of X with 1,000 random permutations of vector y
res <- permutation_correlation_test(X, y, n_perm = 1e3, n_core = 8, method = "spearman")

head(res)

#>              sample.1  sample.2  sample.3   sample.4  sample.5  sample.6
#> feature.1 0.080750138 0.2383798 0.3681659 0.28145490 0.4339816 0.6502498
#> feature.2 0.834333037 0.4116800 0.6117696 0.53045091 0.3954608 0.9099010
#> feature.3 0.600760886 0.7460717 0.4145562 0.43285543 0.5478450 0.2007788
#> feature.4 0.157208442 0.7234506 0.2123919 0.91717922 0.7618364 0.9831965
#> feature.5 0.007399441 0.6145241 0.7327437 0.02747746 0.2540708 0.2980464
#> feature.6 0.466393497 0.4739806 0.4595849 0.82490547 0.5392890 0.7684814
#>            sample.7    sample.8  sample.9  sample.10         cor empirical.p
#> feature.1 0.8687742 0.001840843 0.2196684 0.75355350  0.28484848       0.207
#> feature.2 0.9243132 0.765659131 0.3728954 0.48601704 -0.13939394       0.494
#> feature.3 0.6553841 0.631946311 0.9997498 0.05772175 -0.05454545       0.506
#> feature.4 0.2126957 0.760505739 0.4179180 0.37563486  0.17575758       0.327
#> feature.5 0.1624281 0.777718500 0.2684782 0.14935525  0.10303030       0.398
#> feature.6 0.4733979 0.837271664 0.1775263 0.60446293  0.21212121       0.282
#>                 FDR
#> feature.1 0.5292339
#> feature.2 0.5292339
#> feature.3 0.5292339
#> feature.4 0.5292339
#> feature.5 0.5292339
#> feature.6 0.5292339