Remove principle components from data

Reconstruct data after removing principle components. This function will reconstruct a truncated version of the original data matrix after removing the specified principle components. The source code for this function was stolen from stack exchange

remove_components(x, components = 1, ...)

Arguments

x
data.frame or matrix of original data.

components
numeric vector of components to remove from original data. Default (1)

...
Additional arguments passed to prcomp()

Value

data.frame of values in the original units after removing components

Examples

# Remove first two components from dataset
trunc <- remove_components(USArrests, components = 1:2)

# original data
head(USArrests)
#>            Murder Assault UrbanPop Rape
#> Alabama      13.2     236       58 21.2
#> Alaska       10.0     263       48 44.5
#> Arizona       8.1     294       80 31.0
#> Arkansas      8.8     190       50 19.5
#> California    9.0     276       91 40.6
#> Colorado      7.9     204       78 38.7

# reconstructed data
head(trunc)
#>              Murder  Assault UrbanPop     Rape
#> Alabama    9.984351 170.8348 66.18042 18.62758
#> Alaska     5.322669 169.5595 61.26554 41.13301
#> Arizona    3.321613 171.0435 65.62516 19.90317
#> Arkansas   7.286444 170.7661 65.46754 21.47442
#> California 5.529386 170.4137 64.02358 28.00639
#> Colorado   7.056285 169.9973 62.97728 33.31759