Summarize RNA-seq expression results

Summarize a results data.frame. Return data.frame of counts of up/down/non-DE genes based on log-fold-change and significance values.

summarize_dge(df, fdr_col = "FDR", lfc_col = "logFC", fdr = 0.05, lfc = 0)

Arguments

df

dataframe of results. Must have columns containing significance values and log-fold changes.

fdr_col

Column name of the data.frame containing the significance level values.

lfc_col

Column name of data.frame containing the log-fold change values.

fdr

numeric. FDR or significance value below which genes are considered significant.

lfc

numeric. abs(log-fold change) value above which genes are considered significant.

Value

data.frame with columns describing the count and percentages of up/down/unperturbed genes

Examples

library(coriell)

# Create a simulated results data.frame
# 50 will have logFC > 0
# 50 will have logFC < 0
# 25 of the up-regulated will have significant FDR values (< 0.0001)
# 25 of the down-regulated will have significant FDR values (< 0.0001)
df <- data.frame(feature_id = paste("gene", 1:100,sep = "."),
                 logFC = c(runif(50, min = -8, max = -0.01), 
                           runif(50, min = 0.01, max = 8)),
                 FDR = c(runif(25, min = 1e-12, max = 1e-4),
                         runif(25, min = 0.06, max = 1),
                         runif(25, min = 1e-12, max = 1e-4),
                         runif(25, min = 0.06, max = 1))
)

# view the results data.frame
head(df)
#>   feature_id     logFC          FDR
#> 1     gene.1 -2.678591 7.316693e-06
#> 2     gene.2 -4.503393 1.907321e-05
#> 3     gene.3 -7.267323 4.967141e-05
#> 4     gene.4 -5.471506 4.160380e-05
#> 5     gene.5 -2.677554 5.079279e-05
#> 6     gene.6 -4.631657 3.766164e-05

# summarize results
summarize_dge(df)
#>     Direction  N Percent
#> 1          Up 25      25
#> 2        Down 25      25
#> 3 Unperturbed 50      50