Z-score a dataframe

Convenience function for scaling a dataframe by rows or columns.

zscore_df(df, by = "row")

Arguments

df
data.frame or matrix of numeric columns

by
a character string indicating which axis to scale by. Either "row" or "column". Default ("row")

Value

data.frame with scaled values

Examples

# generate example data
df <- data.frame(
  a = runif(10),
  b = runif(10),
  c = runif(10)
)
rownames(df) <- paste0("row", rownames(df))

# z-score each row of the data.frame
zscore_df(df, by = "row")

#>                a          b           c
#> row1   0.7117193  0.4315994 -1.14331869
#> row2  -0.5546864  1.1544078 -0.59972134
#> row3   1.0390448 -0.9557442 -0.08330065
#> row4  -1.1375518  0.7404791  0.39707271
#> row5  -1.1416380  0.4208291  0.72080895
#> row6   0.7874970  0.3376107 -1.12510775
#> row7  -0.8047020 -0.3148247  1.11952671
#> row8  -0.6389655  1.1524247 -0.51345919
#> row9   0.6563139 -1.1509200  0.49460613
#> row10 -0.8658375 -0.2287036  1.09454106